% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stylecode.R
\name{stylecode}
\alias{stylecode}
\title{Defines the css style for a table}
\usage{
stylecode(name, colh = "#B8B8B8", colh.text = "#111111", col = "#F0F0F0",
  col.text = "#111111", border.col = "gray55", border.wt = 1,
  border.style = "solid", font.size = c(18, 16), lineheight = 1.3,
  hfont.wt = 200, borderh.wt = 2, borderh.col = "#CC0000",
  borderh.style = "none", text.align = "left", htext.align = "left",
  borderb.wt = 1, borderb.col = "#CCCCCC", borderb.style = "none",
  pad = c(3, 3, 12, 3), padh = c(3, 3, 12, 3), collapse = "collapse")
}
\arguments{
\item{name}{The unique id for the style to be referenced by \code{tablecode()}}

\item{colh}{The color of the header}

\item{colh.text}{The color of the text in the header}

\item{col}{The background color}

\item{col.text}{The color of text}

\item{border.col}{Border color}

\item{border.wt}{Border weight - needs to be an integer}

\item{border.style}{The style of border. Options
are \code{none}, \code{hidden},
 \code{solid}, \code{dotted}, \code{double}, \code{dashed},
 \code{groove}, \code{ridge}, \code{inset}, \code{outset}}

\item{font.size}{a vector of font sizes for i) header ii) cells}

\item{lineheight}{The height of rows}

\item{hfont.wt}{The weight of the font in the header
options are \code{normal}, \code{bold}, \code{bolder}, \code{lighter},
or can be a number (usually between 100-900, default is 200).}

\item{borderh.wt}{Border at bottom of header weight - needs to be an
integer}

\item{borderh.col}{Border at bottom of header color}

\item{borderh.style}{The style of border at bottom of header. Options
are same as for \code{border.style}}

\item{text.align}{The alignment of the cell text
are \code{left}, \code{center},
 \code{right}, \code{justify}}

\item{htext.align}{The alignment of the header text
are \code{left}, \code{center},
 \code{right}, \code{justify}}

\item{borderb.wt}{Border weight at bottom of rows}

\item{borderb.col}{Border color at bottom of rows}

\item{borderb.style}{The style of border at the bottom of cells. Options
are same as for \code{border.style}}

\item{pad}{The padding around borders of cells.}

\item{padh}{The padding around borders of header.}

\item{collapse}{The border collapse style - can be
\code{collapse},  \code{separate},  \code{initial}}
}
\value{
A css style
}
\description{
Defines the css style for a table
}
\examples{
stylecode(name="uniqueid")
tablecode(paintings, tabletype="uniqueid", width="80\%")
}

