% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tablecode.R
\name{tablecode}
\alias{tablecode}
\title{Returns a Simple Table in RMarkdown HTML output}
\usage{
tablecode(df, tabletype = "tabletype", width = "80\%")
}
\arguments{
\item{df}{The dataframe to turn into a table}

\item{tabletype}{the name of the table style.  The options currently
available are \code{minimal}, \code{onecol}, \code{box}, \code{hoverTable}, \code{zebra},
   \code{zebra1}, \code{another}, \code{simple}, \code{gridtable}}

\item{width}{the width of the table}
}
\value{
A table of the desired style
}
\description{
Returns a Simple Table in RMarkdown HTML output
}
\examples{
tablecode(paintings, tabletype="minimal")
}

